Minor cleanup of WindowSummarizer. Moved to summarize.py
